challenges/split-text: Add solution to the second part
I just copied and pasted the first part, modified the two last unit tests,
not only to make sure the code works but also to understand all the edge
cases. This has made me realise that I need to worry about words where it
would be easy to think the should be cut with a hyphen, but then the result
would be a line with a useless ending hyphen.

For example, with length=4 and text="is one" the expected output is not
["is -", ...], but rather ["is", "one"]. That's why we need to check, not
only the last letter, but the previous one.