Put together all authentication logic
It shouldn't be together actually, and it should all be testable, but
ReactJS doesn't offer any nice framework for that, so I'd need to use some
tool that gives some Dependency Injection abilities in order to make this
whole thing testable. Anyway, just putting it all together because this is
just a prototype to see the whole thing with Docker-Django-ReactJS work
together.