ReactJS-Django: Don't ignore node_modules
Because we'll need to inject checked out versions from the developer's
machine into the container (although these can be built from the container
itself). I think it's more suitable to "freeze" all dependencies like this
and be able to see with Git if there was any change made to those when
working on new features.